requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
altgraph==0.17.4
    # via
    #   macholib
    #   modulegraph
    #   py2app
macholib==1.16.3
    # via py2app
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
modulegraph==0.19.6
    # via py2app
packaging==24.2
    # via py2app
py2app==0.28.8
    # via -r requirements.in
pygments==2.19.1
    # via rich
pyobjc-core==11.0
    # via pyobjc-framework-cocoa
pyobjc-framework-cocoa==11.0
    # via -r requirements.in
rich==13.9.4
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
